+++ app.js
@@ -26,7 +26,7 @@ function start() {
     const request = require('request')
     const del = require('del')
 
-    let engine = require('./engines/scrapers')
+    let engine = require('./engine/engine')
 
     // mock some services on local
     // if (env === 'development') {
